Zero's firing logic completed
1. Refined the arrow's image
2. Added to the Zero's data file his attacking speed and initial firing accurracy
3. Refined the attacking logic for the playing objects (Zero or Monsters) in a way that the attack starts immediately but not when the attacking speed time has elapsed.
4. Don't set as inactive Zero (for now) as the way the other playing objects are - later there the game over event to be raised
5. The projectiles are rotated to the direction they are moving to
6. Added inaccuracy to the Zero's firing skills - for

+++ MovingCharacter.cpp
@@ -84,7 +84,7 @@ void MovingCharacter::processMonsterCollision(GameObject* monster)
 
 void MovingCharacter::setInsideWindow()
 {
-	float windowW, windowH;
+	double windowW, windowH;
 	extern ResourcesManager *resMan;
 	resMan->getWindowDimensions(windowW, windowH);
 	if (m_rect.x < 0) m_rect.x = 0;
